requirements.txt
streamlit>=1.30.0
reportlab>=4.0
python-barcode>=0.15.1
gspread>=6.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
Pillow>=10.0.0
